daemon: Sanitize the GET /rooms/{room_id}/messages filters as well.
Fractal started to apply filters on their scroll-back requests, this
sadly doesn't include encrypted messages so the scroll-back appeared to
be broken for encrypted rooms.

This patch modifies the filter sanitization logic, since Fractal also
sends out a new filter format that we didn't support, and applies it to
the GET /rooms/{room_id}/messages endpoint as well.