pantalaimon: Prevent the sending of messages if the client didn't sync
Pantalaimon depends on the room state to be present to decide correctly
if a message should be encrypted or now. The room state is fetched from
the server every time we restart, this means that failure to fetch room
states from the server would hinder us from doing the correct decision.

This patch prevents a downgrade to sending unencrypted messages if we're
unable to sync at least once with the server.